final commit

+++ actions/todo.ts
@@ -2,11 +2,26 @@
 
 import { revalidatePath } from "next/cache";
 import prisma from "@/db";
+import { getUser } from "./user";
 
-export async function addTodo(todoTitle: string, expiryDate: Date | string) {
+
+export async function addTodo(
+      todoTitle: string,
+      expiryDate: Date | string,
+      userEmail: string
+) {
       try {
             if (!todoTitle || !expiryDate) {
-                  return { error: "Title and expiry date are required", success: false };
+                  return {
+                        error: "Title and expiry date are required",
+                        success: false,
+                  };
+            }
+
+            const { user } = await getUser(userEmail);
+
+            if (!user) {
+                  return { error: "User not found", success: false };
             }
 
             const todo = await prisma.todo.create({
@@ -16,25 +31,38 @@ export async function addTodo(todoTitle: string, expiryDate: Date | string) {
                         completed: false,
                         expired: false,
                         createdAt: new Date(),
+                        userId: user?.id,
                   },
             });
 
             // This tells Next.js to revalidate the cache for the /todos route
             revalidatePath("/todos");
-            
-            return { todo, success: true };
 
+            return { todo, success: true };
       } catch (error) {
             console.error("Failed to create todo:", error);
             return { error: "Failed to create todo", success: false };
       }
 }
 
-export async function getTodos() {
+export async function getTodos(userEmail: string) {
       try {
+            if (!userEmail) {
+                  return { error: "User email is required", success: false };
+            }
+
+            const { user } = await getUser(userEmail);
+
+            if (!user) {
+                  return { error: "User not found", success: false };
+            }
+
             const todos = await prisma.todo.findMany({
                   orderBy: {
                         createdAt: "desc",
+                  },
+                  where: {
+                        userId: user?.id,
                   },
             });
             return { todos, success: true };
@@ -44,11 +72,21 @@ export async function getTodos() {
       }
 }
 
-export async function updateTodo(id: number, data: { title?: string; expiresAt?: Date; completed?: boolean }) {
+export async function updateTodo(
+      id: number,
+      userEmail: string,
+      data: { title?: string; expiresAt?: Date; completed?: boolean }
+) {
       try {
+            const { user } = await getUser(userEmail);
+
+            if (!user) {
+                  return { error: "User not found", success: false };
+            }
+
             const todo = await prisma.todo.update({
-                  where: { id },
-                  data
+                  where: { id, userId: user.id },
+                  data,
             });
             revalidatePath("/todos");
             return { todo, success: true };
@@ -58,14 +96,20 @@ export async function updateTodo(id: number, data: { title?: string; expiresAt?:
       }
 }
 
-export async function deleteTodo(id: number) {
+export async function deleteTodo(id: number, userEmail: string) {
       try {
             if (!id) {
                   return { error: "Todo ID is required", success: false };
             }
 
+            const { user } = await getUser(userEmail);
+
+            if (!user) {
+                  return { error: "User not found", success: false };
+            }
+
             const todo = await prisma.todo.delete({
-                  where: { id }
+                  where: { id, userId: user.id },
             });
             revalidatePath("/todos");
             return { todo, success: true };
